talentVerify_backend/requirements.txt
asgiref==3.7.2
diff-match-patch==20230430
Django==5.0.3
django-cors-headers==4.3.1
django-import-export==4.0.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2024.1
